Rename lib subdirectory so as not to conflict with ruby-debug
The namespaces still conflict, but now at least you can have
both ruby-debug and ruby-debug-ide on the load path without
blowing up, as long as you require only one of them. For example,
if you have both ruby-debug and ruby-debug-ide in your Gemfile
with :require => false.

Plus it's just good practice to name the lib subdirectory the same
as the gem. And now we no longer need to conditionally use
require_relative.
