+++ lib/ruby-debug-ide/commands/catchpoint.rb
