+++ lib/ruby-debug-ide/commands/condition.rb
