+++ lib/ruby-debug-ide/commands/control.rb
