+++ lib/ruby-debug-ide/commands/enable.rb
