+++ lib/ruby-debug-ide/commands/eval.rb
