+++ lib/ruby-debug-ide/commands/frame.rb
