+++ lib/ruby-debug-ide/commands/inspect.rb
