+++ lib/ruby-debug-ide/commands/pause.rb
