+++ lib/ruby-debug-ide/commands/set_type.rb
