+++ lib/ruby-debug-ide/commands/threads.rb
