+++ lib/ruby-debug-ide/commands/variables.rb
