+++ lib/ruby-debug-ide/helper.rb
