+++ lib/ruby-debug-ide/xml_printer.rb
